requirements.txt
playwright>=1.44.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
rich>=13.7.0
typer>=0.12.0
python-dotenv>=1.0.0
openai>=1.30.0
tldextract>=5.1.0
urllib3>=2.2.0
dnspython>=2.6.0
robotexclusionrulesparser>=1.7.1
